+++ common/src/main/java/de/maxhenkel/tradecycling/TradeCyclingMod.java
@@ -1,13 +1,12 @@
 package de.maxhenkel.tradecycling;
 
-import de.maxhenkel.tradecycling.mixin.AbstractVillagerAccessor;
+import de.maxhenkel.tradecycling.mixin.VillagerAccessor;
 import de.maxhenkel.tradecycling.mixin.MerchantMenuAccessor;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.entity.npc.Villager;
 import net.minecraft.world.inventory.MerchantMenu;
 import net.minecraft.world.item.trading.Merchant;
-import net.minecraft.world.item.trading.MerchantOffers;
 
 public abstract class TradeCyclingMod {
 
@@ -34,12 +33,13 @@ public abstract class TradeCyclingMod {
         if (!(merchant instanceof Villager villager)) {
             return;
         }
-        if (!(merchant instanceof AbstractVillagerAccessor villagerAccessor)) {
+        if (!(merchant instanceof VillagerAccessor villagerAccessor)) {
             return;
         }
 
-        villagerAccessor.setOffers(new MerchantOffers());
-        villagerAccessor.invokeUpdateTrades();
+        villager.setOffers(null);
+        villagerAccessor.invokeUpdateSpecialPrices(player);
+        villager.setTradingPlayer(player);
         player.sendMerchantOffers(container.containerId, villager.getOffers(), villager.getVillagerData().getLevel(), villager.getVillagerXp(), villager.showProgressBar(), villager.canRestock());
     }